query attempted
we have a drop-down menu on a static page to choose from now, but it only works once. so if we want to find data for another country we have to run node index.js again.

also, did not make any changes to /graph route, so it should be just as it was before, but I didn't visit the page.